favorites
finished favorites cards. added favorites filters. REMOVED react headroom for header pop up.

+++ src/pages/favorites.js
@@ -1,77 +1,13 @@
 import { useState, useEffect } from "react";
-import Link from "next/link";
 
-const baseUrlImg = "https://image.tmdb.org/t/p/w500/";
+import { MovieCard, TvCard, PplCard } from "@/components/FavoriteCards";
 
-const MovieCard = ({ data }) => (
-  <>
-    <img
-      className="w-full"
-      src={baseUrlImg + data.poster_path}
-      alt={data.title}
-    />
-    <div>
-      <div className="text-center font-bold text-xl mb-2">{data.title}</div>
-      <p className="text-gray-700 text-base text-justify line-clamp-4">{data.overview}</p>
-    </div>
-    <div className="px-6 py-4">
-      <Link
-        href={`/movie/${data.id}`}
-        className="text-blue-500 hover:underline"
-      >
-        Details
-      </Link>
-    </div>
-  </>
-);
-
-const TVCard = ({ data }) => (
-  <>
-    <img
-      className="w-full"
-      src={baseUrlImg + data.poster_path}
-      alt={data.name}
-    />
-    <div>
-      <div className="text-center font-bold text-xl mb-2">{data.name}</div>
-      <p className="text-gray-700 text-base text-justify line-clamp-4">
-        {data.overview}
-      </p>
-    </div>
-    <div className="px-6 py-4">
-      <Link href={`/tv/${data.id}`} className="text-blue-500 hover:underline">
-        Details
-      </Link>
-    </div>
-  </>
-);
-
-const PeopleCard = ({ data }) => (
-  <>
-    <img
-      className="w-full"
-      src={baseUrlImg + data.profile_path}
-      alt={data.name}
-    />
-    <div>
-      <div className="text-center font-bold text-xl mb-2">{data.name}</div>
-      <p className="text-gray-700 text-base text-justify line-clamp-4">
-        {data.biography}
-      </p>
-    </div>
-    <div className="px-6 py-4">
-      <Link
-        href={`/people/${data.id}`}
-        className="text-blue-500 hover:underline"
-      >
-        Details
-      </Link>
-    </div>
-  </>
-);
+import NotFound from "@/components/notFound";
 
 export default function Favorites() {
   const [favoriteFilms, setFavoriteFilms] = useState([]);
+  const [filter, setFilter] = useState([]);
+  const [search, setSearch] = useState("");
 
   useEffect(() => {
     const favoritesJSON = localStorage.getItem("favorites");
@@ -81,11 +17,27 @@ export default function Favorites() {
   }, []);
 
   const toggleFavorite = (entityId, entityType) => {
-    const updatedFavorites = favoriteFilms.filter(
-      (entity) => !(entity.data.id === entityId && entity.type === entityType)
+    const confirmed = window.confirm(
+      "Are you sure you want to remove this item from your favorites?"
     );
-    setFavoriteFilms(updatedFavorites);
-    localStorage.setItem("favorites", JSON.stringify(updatedFavorites));
+
+    if (confirmed) {
+      const updatedFavorites = favoriteFilms.filter(
+        (entity) => !(entity.data.id === entityId && entity.type === entityType)
+      );
+      setFavoriteFilms(updatedFavorites);
+      localStorage.setItem("favorites", JSON.stringify(updatedFavorites));
+    }
+  };
+
+  const toggleFilter = (filterType) => {
+    filter.includes(filterType)
+      ? setFilter(filter.filter((item) => item !== filterType))
+      : setFilter([...filter, filterType]);
+  };
+
+  const handleSearchChange = (e) => {
+    setSearch(e.target.value);
   };
 
   if (!(favoriteFilms.length > 0)) {
@@ -103,26 +55,111 @@ export default function Favorites() {
     );
   }
 
+  const filteredData = favoriteFilms.filter((element) => {
+    if (filter.length == 0) {
+      return true;
+    } else {
+      return filter.includes(element.type);
+    }
+  })
+
+  const searchedData = filteredData.filter((element) => {
+    if (element.data.title) {
+      return element.data.title.toLowerCase().includes(search.toLowerCase());
+    } else {
+      return element.data.name.toLowerCase().includes(search.toLowerCase());
+    }
+  });
+
   return (
-  <div className="flex flex-wrap justify-around">
-      {
-        favoriteFilms.map((entity) => (
-        <div
-          key={entity.data.id}
-          className="max-w-xs overflow-hidden shadow-lg p-6 mt-4 bg-gray-200"
-        >
-          {entity.type === "movie" && <MovieCard data={entity.data} />}
-          {entity.type === "tv" && <TVCard data={entity.data} />}
-          {entity.type === "people" && <PeopleCard data={entity.data} />}
-          <button
-            className="bg-red-400 text-white px-4 py-2 rounded-lg hover:bg-red-600"
-            onClick={() => toggleFavorite(entity.data.id, entity.type)}
-          >
-            Remove
-          </button>
+    <>
+      <div className="flex flex-wrap items-center -mx-4 space-x-8 text-xl justify-center text-gray-100">
+        <div className="flex overflow-hidden rounded-md text-gray-100">
+          <input
+            type="search"
+            name="search"
+            placeholder="Search..."
+            className={`bg-transparent focus:outline-none text-gray-800 p-3 w-full h-full border-b-4 ${
+              search.length > 0 ? "border-violet-400" : "border-gray-700"
+            }`}
+            onChange={(e) => {
+              handleSearchChange(e);
+            }}
+          />
+          <div className="flex items-center justify-center px-4 bg-gray-700 text-gray-800">
+            <svg
+              fill="currentColor"
+              viewBox="0 0 512 512"
+              className="w-6 h-6 text-gray-100"
+            >
+              <path d="M479.6,399.716l-81.084-81.084-62.368-25.767A175.014,175.014,0,0,0,368,192c0-97.047-78.953-176-176-176S16,94.953,16,192,94.953,368,192,368a175.034,175.034,0,0,0,101.619-32.377l25.7,62.2L400.4,478.911a56,56,0,1,0,79.2-79.195ZM48,192c0-79.4,64.6-144,144-144s144,64.6,144,144S271.4,336,192,336,48,271.4,48,192ZM456.971,456.284a24.028,24.028,0,0,1-33.942,0l-76.572-76.572-23.894-57.835L380.4,345.771l76.573,76.572A24.028,24.028,0,0,1,456.971,456.284Z"></path>
+            </svg>
+          </div>
         </div>
-        ))
-      }
-    </div>
-  )
+        <button
+          rel="noopener noreferrer"
+          href="#"
+          className={`flex items-center flex-shrink-0 px-5 py-2 border-b-4 text-gray-400 ${
+            filter.includes("movie") ? "border-violet-400" : "border-gray-700"
+          }`}
+          onClick={() => toggleFilter("movie")}
+        >
+          Movie
+        </button>
+        <button
+          rel="noopener noreferrer"
+          href="#"
+          className={`flex items-center flex-shrink-0 px-5 py-2 border-b-4 text-gray-400 ${
+            filter.includes("tv") ? "border-violet-400" : "border-gray-700"
+          }`}
+          onClick={() => toggleFilter("tv")}
+        >
+          TV show
+        </button>
+        <button
+          rel="noopener noreferrer"
+          href="#"
+          className={`flex items-center flex-shrink-0 px-5 py-2 border-b-4 text-gray-400 ${
+            filter.includes("people") ? "border-violet-400" : "border-gray-700"
+          }`}
+          onClick={() => toggleFilter("people")}
+        >
+          People
+        </button>
+      </div>
+      {!(searchedData.length > 0) ? (
+        <div className="max-w-lg mx-auto">
+          <NotFound />
+        </div>
+      ) : (
+        <div className="flex flex-wrap gap-8 mt-12">
+          {searchedData.map((entity) => (
+            <>
+              {entity.type === "movie" && (
+                <MovieCard
+                  data={entity.data}
+                  type={entity.type}
+                  toggleFavorite={toggleFavorite}
+                />
+              )}
+              {entity.type === "tv" && (
+                <TvCard
+                  data={entity.data}
+                  type={entity.type}
+                  toggleFavorite={toggleFavorite}
+                />
+              )}
+              {entity.type === "people" && (
+                <PplCard
+                  data={entity.data}
+                  type={entity.type}
+                  toggleFavorite={toggleFavorite}
+                />
+              )}
+            </>
+          ))}
+        </div>
+      )}
+    </>
+  );  
 }
